add middleware

+++ backend/internal/delivery/http/post/delete.go
@@ -4,6 +4,7 @@ import (
 	"backend/pkg/exceptions"
 	"backend/pkg/utils"
 	"context"
+	"errors"
 	"net/http"
 
 	"github.com/gorilla/mux"
@@ -12,6 +13,17 @@ import (
 func (x *postHttpHandler) Delete(w http.ResponseWriter, r *http.Request) {
 	vars := mux.Vars(r)
 	id := vars["id"]
+
+	_, role, errJwt := utils.DecodeJwtToken(r)
+	if errJwt != nil {
+		utils.RespondWithError(w, http.StatusBadRequest, []error{errJwt})
+		return
+	}
+
+	if role != "admin" && role != "constributor" {
+		utils.RespondWithError(w, http.StatusUnauthorized, []error{errors.New("not authorization")})
+		return
+	}
 
 	err := x.postUsecase.Delete(context.Background(), id)
